Moved hooks into multiple cpp files into `Hooks` directory
