+++ YAGA/YAGA.vcxproj
@@ -938,6 +938,13 @@
       </SubType>
     </ClInclude>
     <ClInclude Include="Plugin.cpp" />
+    <ClInclude Include="Hooks\oCAIHuman.cpp" />
+    <ClInclude Include="Hooks\oCItemContainer.cpp" />
+    <ClInclude Include="Hooks\oCMobInter.cpp" />
+    <ClInclude Include="Hooks\oCNpc.cpp" />
+    <ClInclude Include="Hooks\zCMenu.cpp" />
+    <ClInclude Include="Hooks\zCViewDialogChoice.cpp" />
+    <ClInclude Include="Hooks\zCVob.cpp" />
     <ClCompile Include="Interface.cpp" />
     <ClCompile Include="ZenGin\Gothic_II_Addon\ClassDeclarators_G2A.cpp" />
     <ClCompile Include="ZenGin\Gothic_II_Addon\Globals_G2A.cpp" />
